+++ DatadogTrace/Sources/Trace.swift
@@ -74,7 +74,8 @@ public enum Trace {
                 firstPartyHosts: firstPartyHosts,
                 traceContextInjection: traceContextInjection,
                 telemetry: core.telemetry,
-                redactedStatusCodes: urlSessionTracking.redactedStatusCodes
+                redactedStatusCodes: urlSessionTracking.redactedStatusCodes,
+                sessionSampling: { [weak core] in core?.rumSessionSampling }
             )
 
             try core.register(urlSessionHandler: urlSessionHandler)